fix: removed font parameter and changed int pairs to float pairs for size and position parameters

+++ gui/UI/shared/UIBlocks/IUIBlock.hpp
@@ -19,7 +19,7 @@ namespace UIBlocks {
         public:
         virtual void draw(zappyGUI::Window &window) = 0;
         virtual void handleEvent(const sf::Event &event) = 0;
-        virtual void setPosition(const std::pair<int, int> &position) = 0;
+        virtual void setPosition(const std::pair<float, float> &position) = 0;
         virtual const std::variant<std::string, std::vector<std::shared_ptr<IUIBlock>>> getValue() const = 0;
     };
 }
